Move LinearExpansion to QCModel/CI.
