--- gqcp/include/ONVBasis/LinearExpansion/CMakeLists.txt
@@ -1,6 +1,0 @@
-target_sources(gqcp
-    PRIVATE
-        SpinUnresolvedLinearExpansion.hpp
-        LinearExpansion.hpp
-        LinearExpansionReader.hpp
-)
