+++ gqcp/include/QCMethod/Geminals/AP1roGGeminalCoefficients.hpp
@@ -18,7 +18,7 @@
 #pragma once
 
 
-#include "ONVBasis/LinearExpansion/LinearExpansion.hpp"
+#include "QCModel/CI/LinearExpansion.hpp"
 #include "Mathematical/Representation/BlockMatrix.hpp"
 #include "Operator/SecondQuantized/SQHamiltonian.hpp"
 #include "QCMethod/Geminals/GeminalCoefficientsInterface.hpp"
