+++ gqcp/include/QCMethod/Geminals/APIGGeminalCoefficients.hpp
@@ -18,7 +18,7 @@
 #pragma once
 
 
-#include "ONVBasis/LinearExpansion/LinearExpansion.hpp"
+#include "QCModel/CI/LinearExpansion.hpp"
 #include "Mathematical/Representation/Matrix.hpp"
 #include "QCMethod/Geminals/GeminalCoefficientsInterface.hpp"
