+++ gqcp/include/QCModel/CI/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(gqcp
+    PRIVATE
+        SpinUnresolvedLinearExpansion.hpp
+        LinearExpansion.hpp
+        LinearExpansionReader.hpp
+)
