+++ gqcp/src/CMakeLists.txt
@@ -5,4 +5,5 @@ add_subdirectory(Molecule)
 add_subdirectory(Operator)
 add_subdirectory(Processing)
 add_subdirectory(QCMethod)
+add_subdirectory(QCModel)
 add_subdirectory(Utilities)
