+++ gqcp/src/ONVBasis/CMakeLists.txt
@@ -9,5 +9,3 @@ target_sources(gqcp
         ProductONVBasis.cpp
         SelectedONVBasis.cpp
 )
-
-add_subdirectory(LinearExpansion)
