+++ gqcp/src/QCModel/CI/CMakeLists.txt
@@ -1,6 +1,6 @@
 target_sources(gqcp
     PRIVATE
-        SpinUnresolvedLinearExpansion.cpp
         LinearExpansion.cpp
         LinearExpansionReader.cpp
+        SpinUnresolvedLinearExpansion.cpp
 )
