+++ gqcp/src/QCModel/CI/LinearExpansion.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-gqcp.  If not, see <http://www.gnu.org/licenses/>.
 // 
-#include "ONVBasis/LinearExpansion/LinearExpansion.hpp"
+#include "QCModel/CI/LinearExpansion.hpp"
 
 #include "ONVBasis/ProductONVBasis.hpp"
 #include "Mathematical/Representation/SquareMatrix.hpp"
