+++ gqcp/src/QCModel/CI/LinearExpansionReader.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-gqcp.  If not, see <http://www.gnu.org/licenses/>.
 // 
-#include "ONVBasis/LinearExpansion/LinearExpansionReader.hpp"
+#include "QCModel/CI/LinearExpansionReader.hpp"
 
 #include <boost/algorithm/string.hpp>
 #include <boost/dynamic_bitset.hpp>
