+++ gqcp/src/QCModel/CI/SpinUnresolvedLinearExpansion.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-gqcp.  If not, see <http://www.gnu.org/licenses/>.
 // 
-#include "ONVBasis/LinearExpansion/SpinUnresolvedLinearExpansion.hpp"
+#include "QCModel/CI/SpinUnresolvedLinearExpansion.hpp"
 
 
 namespace GQCP {
