+++ gqcp/tests/ONVBasis/SelectedONVBasis_test.cpp
@@ -22,7 +22,7 @@
 #include "ONVBasis/SelectedONVBasis.hpp"
 
 #include "Basis/transform.hpp"
-#include "ONVBasis/LinearExpansion/LinearExpansionReader.hpp"
+#include "QCModel/CI/LinearExpansionReader.hpp"
 
 
 /**
