+++ gqcp/tests/QCModel/CI/LinearExpansion_test.cpp
@@ -19,7 +19,7 @@
 
 #include <boost/test/unit_test.hpp>
 
-#include "ONVBasis/LinearExpansion/LinearExpansion.hpp"
+#include "QCModel/CI/LinearExpansion.hpp"
 
 #include "Basis/transform.hpp"
 #include "ONVBasis/ONVBasis.hpp"
