+++ gqcp/tests/QCModel/CMakeLists.txt
@@ -1,3 +1,4 @@
+add_subdirectory(CI)
 add_subdirectory(HF)
 
 set(test_target_sources ${test_target_sources} PARENT_SCOPE)
